@@ -1,7 +1,11 @@
 Airbrake.configure do |config|
-  config.api_key = Rails.application.secrets.airbrake["api_key"]
-  config.host    = Rails.application.secrets.airbrake["host"]
-  config.port    = Rails.application.secrets.airbrake["port"]
-  config.secure  = config.port == 443
-end
+  airbrake_secrets = Rails.application.secrets.airbrake
 
+  protocol = "http#{airbrake_secrets['port'].to_i == 443 ? 's' : ''}"
+  host = airbrake_secrets['host']
+  port = airbrake_secrets['port']
+
+  config.project_key = airbrake_secrets["api_key"]
+  config.host = "#{protocol}://#{host}:#{port}"
+  config.project_id = 1
+end
