@@ -1,9 +1,9 @@
-::SecureHeaders::Configuration.configure do |config|
-  config.hsts = {:max_age => 20.years.to_i, :include_subdomains => false}
+::SecureHeaders::Configuration.default do |config|
+  config.hsts = "max-age=#{20.years.to_i}"
   config.x_frame_options = 'SAMEORIGIN'
   config.x_content_type_options = "nosniff"
-  config.x_xss_protection = {:value => 1, :mode => 'block'}
+  config.x_xss_protection = '1; mode=block'
   config.x_download_options = 'noopen'
   config.x_permitted_cross_domain_policies = 'none'
-  config.csp = false
+  config.csp = SecureHeaders::OPT_OUT
 end
