@@ -1,35 +1,32 @@
 require 'test_helper'
+require 'integration/concerns/login_helpers'
 
 class PasswordsIntegrationTest < ActionDispatch::IntegrationTest
+  include Participa::Test::LoginHelpers
 
   setup do
     @user = FactoryGirl.create(:user)
     @legacy_password_user = FactoryGirl.create(:user, :legacy_password_user)
   end
 
-  def login user
-    post_via_redirect user_session_path, 'user[login]' => user.email, 'user[password]' => user.password 
-  end
-
   test "should login with password as user" do
     login @user
-    get '/es'
-    assert_response :success
+    assert_title 'No te has verificado'
+    assert_equal root_path(locale: 'es'), current_path
   end
 
   test "should login with legacy password and change it as legacy_password_user" do
     login @legacy_password_user
-    get '/es'
-    assert_response :redirect
-    assert_redirected_to new_legacy_password_path
+    assert_equal new_legacy_password_path(locale: 'es'), current_path
   end
 
   test "should not have legacy password if legacy_password_user changes it through devise" do
     password = 'lalalilo'
-    put "/es/users/password", user: {reset_password_token: @legacy_password_user.reset_password_token, password: password, password_confirmation: password} 
-    post_via_redirect user_session_path, 'user[email]' => @legacy_password_user.email, 'user[password]' => password 
-    get '/es'
-    assert_response :success
+    reset_password_token = @legacy_password_user.send_reset_password_instructions
+    put "/es/users/password", params: { user: {reset_password_token: reset_password_token, password: password, password_confirmation: password} } 
+    login @legacy_password_user, password
+    assert_title 'No te has verificado'
+    assert_equal root_path(locale: 'es'), current_path
   end
 
 end
