@@ -15,7 +15,7 @@ include Warden::Test::Helpers
 Warden.test_mode!
 
 class ActionController::TestCase
-  include Devise::TestHelpers
+  include Devise::Test::ControllerHelpers
   include FactoryGirl::Syntax::Methods
 end
 
